Changes the fugutive merge, default to signing
The fugutive merge is now set to use terminal nvim, not nvim-qt, because
that didn't recognise the options, apparently, and it's better to have
it working. I also switched signing commits by default on for now - not
sure I'll keep it because I'm not sure there's a point to it, but hey.